requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.4.0
torchvision
numpy
pillow
torchdiffeq==0.1
pandas
seaborn
munch
pyyaml
pytest
tensorboard
tensorflow
azureml-sdk[tensorboard]